backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
pytz==2023.2
sqlparse==0.4.3
typing_extensions==4.5.0
Pillow==9.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-cors-headers==3.14.0
django-filter==23.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.6.0
drf-extra-fields==3.4.1
python-dotenv==0.19.0
reportlab==3.6.12
